[Refactor] 챌린지 기록 평가 항목 변경 (#319)
* refactor: 기획 변경에 따른 DTO 평가 타입 변경

* refactor: imgUrl 경로 예시 추가

* refactor: 챌린지 지출 평가 변경된 기획 반영

* refactor: EmojiType 패키지 경로 변경 (record 밑에서 emoji 밑으로)

* refactor: 사용하지 않는 메서드 삭제
