+++ Module-API/src/main/java/depromeet/api/domain/record/dto/response/CreateRecordResponse.java
@@ -17,7 +17,7 @@ public class CreateRecordResponse {
     private String title;
     private String content;
     private String imgUrl;
-    private Integer evaluation;
+    private String evaluation;
     private RecordUserInfo userInfo;
 
     public static CreateRecordResponse of(Record record) {
