+++ Module-API/src/main/java/depromeet/api/domain/record/mapper/RecordMapper.java
@@ -6,7 +6,6 @@ import depromeet.api.domain.record.dto.response.CreateRecordResponse;
 import depromeet.api.domain.record.dto.response.GetRecordResponse;
 import depromeet.common.annotation.Mapper;
 import depromeet.domain.challenge.domain.Challenge;
-import depromeet.domain.record.domain.Evaluation;
 import depromeet.domain.record.domain.Record;
 import depromeet.domain.user.domain.User;
 import depromeet.domain.userchallenge.domain.UserChallenge;
@@ -29,7 +28,7 @@ public class RecordMapper {
                 createRecordRequest.getTitle(),
                 createRecordRequest.getContent(),
                 createRecordRequest.getImgUrl(),
-                Evaluation.getEnumTypeByValue(createRecordRequest.getEvaluation()));
+                createRecordRequest.getEvaluation());
     }
 
     public CreateRecordResponse toCreateRecordResponse(Record record) {
