+++ Module-API/src/test/java/depromeet/api/domain/record/controller/RecordControllerTest.java
@@ -19,6 +19,7 @@ import depromeet.api.domain.record.usecase.DeleteRecordUseCase;
 import depromeet.api.domain.record.usecase.GetRecordUseCase;
 import depromeet.api.domain.record.usecase.UpdateRecordUseCase;
 import depromeet.api.util.AuthenticationUtil;
+import depromeet.domain.record.domain.Evaluation;
 import java.util.Objects;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
@@ -86,7 +87,7 @@ class RecordControllerTest {
                         .title("커피")
                         .content("커피는 무죄야")
                         .imgUrl("")
-                        .evaluation(1)
+                        .evaluation(Evaluation.CRAZY)
                         .build();
 
         CreateRecordResponse createRecordResponse =
@@ -95,7 +96,7 @@ class RecordControllerTest {
                         .title("커피")
                         .content("커피는 무죄야")
                         .imgUrl("")
-                        .evaluation(1)
+                        .evaluation("CRAZY")
                         .build();
 
         MockHttpServletRequestBuilder requestBuilder =
@@ -133,7 +134,7 @@ class RecordControllerTest {
                         .title(" ")
                         .content("커피는 무죄야")
                         .imgUrl("")
-                        .evaluation(1)
+                        .evaluation(Evaluation.CRAZY)
                         .build();
 
         MockHttpServletRequestBuilder requestBuilder =
@@ -168,7 +169,7 @@ class RecordControllerTest {
                         .title("커피")
                         .content("커피는 맛있어")
                         .imgUrl("")
-                        .evaluation(1)
+                        .evaluation(Evaluation.CRAZY)
                         .build();
 
         CreateRecordResponse createRecordResponse =
@@ -177,7 +178,7 @@ class RecordControllerTest {
                         .title("커피")
                         .content("커피는 무죄야")
                         .imgUrl("")
-                        .evaluation(1)
+                        .evaluation("CRAZY")
                         .build();
 
         MockHttpServletRequestBuilder requestBuilder =
@@ -206,7 +207,7 @@ class RecordControllerTest {
                         .title("커피")
                         .content("커피는 무죄야")
                         .imgUrl("")
-                        .evaluation(1)
+                        .evaluation("CRAZY")
                         .build();
 
         MockHttpServletRequestBuilder requestBuilder =
