+++ Module-Domain/src/main/java/depromeet/domain/record/domain/Evaluation.java
@@ -1,25 +1,15 @@
 package depromeet.domain.record.domain;
 
 
-import depromeet.domain.record.exception.RecordEvaluationNotFoundException;
-import java.util.Arrays;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
 public enum Evaluation {
-    ONE(1),
-    TWO(2),
-    THREE(3),
-    FOUR(4);
+    WELLDONE("WELLDONE"),
+    REGRETFUL("REGRETFUL"),
+    CRAZY("CRAZY");
 
-    private final int value;
-
-    public static Evaluation getEnumTypeByValue(int value) {
-        return Arrays.stream(Evaluation.values())
-                .filter(type -> type.getValue() == value)
-                .findFirst()
-                .orElseThrow(() -> RecordEvaluationNotFoundException.EXCEPTION);
-    }
+    private final String value;
 }
